### 2025-01-29 #7737 @editorialbot
[PRE REVIEW]: REMORA: Regional Modeling of Oceans Refined Adptively (built on AMReX)
**Submitting author:** @hklion (<a href="http://orcid.org/0000-0003-2095-4293">Hannah Klion</a>)
**Repository:** https://github.com/seahorce-scidac/REMORA
**Branch with paper.md** (empty if default branch):
**Version:** v1.0
