### 2025-01-30 #7737 @hklion
Thanks @kthyng! A few suggestions for reviewers:

- parkermac
- jcwarner-usgs
- glwagner
- adcroft
- markusReinert
- JessCG (COI note: In summer/fall 2023, we had a few meetings to discuss how Jessica and others at Oregon State might be able to use REMORA (then ROMS-X) when it was more mature. When building REMORA, Jessica corrected a directory name in our documentation [(PR)](https://github.com/seahorce-scidac/REMORA/pull/38), so she appears on the commit author list. We have not had any contact since fall 2023. I don't believe there's a conflict here, but still wanted to flag it.)
