### 2025-01-30 #7737 @kthyng
@hklion Thanks! A few notes on your paper:
* Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example, "european" is not capitalized in one reference, but please check thoroughly for others.
* Please check that your inline vs. parenthetical references are correct. You can compare with the example paper here: https://joss.readthedocs.io/en/latest/example_paper.html
