### 2025-01-30 #7737 @hklion
@kthyng Thanks for the examples. I've updated the citations so the formatting and capitalization are more consistent.
