### 2025-02-20 5cfd6cd
initial bulk flux commits, plus add things to constants file
