+++ Exec/BoundaryLayer/prob.cpp
@@ -19,8 +19,8 @@ amrex_probinit(
  * \brief Initializes bathymetry h and surface height Zeta
  */
 void
-init_custom_bathymetry (int /*lev*/, const Geometry& geom,
-                        MultiFab& mf_h, REMORA const& /*remora*/,
+init_custom_bathymetry (int lev, const Geometry& geom,
+                        MultiFab& mf_h, REMORA const& remora,
                         const SolverChoice& m_solverChoice,
                         int /*rrx*/, int /*rry*/)
 {
@@ -34,12 +34,14 @@ init_custom_bathymetry (int /*lev*/, const Geometry& geom,
 
     mf_h.setVal(geomdata.ProbHi(2));
 
-    const int Lm = geom.Domain().size()[0];
-    const int Mm = geom.Domain().size()[1];
+    const int Lm = geomdata.Domain().size()[0];
+    const int Mm = geomdata.Domain().size()[1];
 
     for ( MFIter mfi(mf_h, TilingIfNotGPU()); mfi.isValid(); ++mfi )
     {
       Array4<Real> const& h  = (mf_h).array(mfi);
+      Array4<const Real> const& xr  = remora.vec_xr[lev]->array(mfi);
+      Array4<const Real> const& yr  = remora.vec_yr[lev]->array(mfi);
 
       Box bx = mfi.tilebox();
       Box gbx2 = bx;
@@ -50,38 +52,42 @@ init_custom_bathymetry (int /*lev*/, const Geometry& geom,
 
       Gpu::streamSynchronize();
 
-      if (NSPeriodic) {
+      ParallelFor(gbx2D, [=] AMREX_GPU_DEVICE (int i, int j, int )
+      {
+          Real val1 = (xr(i,j,0) + 500.0_rt) / 15000.0_rt;
+          h(i,j,0) = 14.0_rt + 25.0_rt * (1.0_rt - std::exp(-PI * xr(i,j,0) * 1.0e-5_rt)) - 8.0_rt * std::exp(-val1 * val1);
+      });
 
-          ParallelFor(gbx2D, [=] AMREX_GPU_DEVICE (int i, int j, int )
-          {
-              int iFort = i+1; // (+1 is to match the Fortran indexing in ROMS)
-
-              Real val1 = (iFort <= Lm/2.0) ? iFort : Lm+1-iFort;
-              val1 -= 0.5;
-              Real adj = geomdata.CellSize()[1]/1000.0_rt;
-
-              h(i,j,0) = std::min(-geomdata.ProbLo(2),(84.5_rt+66.526_rt*std::tanh((val1*adj-10.0_rt)/7.0_rt)));
-          });
-
-      } else if (EWPeriodic) {
-
-          ParallelFor(gbx2D, [=] AMREX_GPU_DEVICE (int i, int j, int )
-          {
-              int jFort = j+1; // (+1 is to match the Fortran indexing in ROMS)
-
-              Real val1 = (jFort<=Mm/2.0) ? jFort : Mm+1-jFort;
-              val1 -= 0.5;
-              Real adj = geomdata.CellSize()[0]/1000.0_rt;
-
-              h(i,j,0) = std::min(-geomdata.ProbLo(2),(84.5_rt+66.526_rt*std::tanh((val1*adj-10.0_rt)/7.0_rt)));
-          });
-      }
     } // mfi
 }
 
 void
-init_custom_grid_scale (int /*lev*/, const Geometry& /*geom*/,
-                   MultiFab& /*mf_pm*/, MultiFab& /*mf_pn*/) {}
[…]
