+++ Source/TimeIntegration/REMORA_rho_eos.cpp
@@ -25,6 +25,8 @@ REMORA::rho_eos (const Box& bx,
                 const Array4<Real      >& rhoA,
                 const Array4<Real      >& rhoS,
                 const Array4<Real      >& bvf,
+                const Array4<Real      >& alpha,
+                const Array4<Real      >& beta,
                 const Array4<Real const>& Hz,
                 const Array4<Real const>& z_w,
                 const Array4<Real const>& z_r,
@@ -35,7 +37,7 @@ REMORA::rho_eos (const Box& bx,
     if (solverChoice.eos_type == EOSType::linear) {
         lin_eos(bx, state, rho, rhoA, rhoS, bvf, Hz, z_w, z_r, h, mskr, N);
     } else if (solverChoice.eos_type == EOSType::nonlinear) {
-        nonlin_eos(bx, state, rho, rhoA, rhoS, bvf, Hz, z_w, z_r, h, mskr, N);
+        nonlin_eos(bx, state, rho, rhoA, rhoS, bvf, alpha, beta, Hz, z_w, z_r, h, mskr, N);
     } else {
         Abort("Unknown EOS type in rho_eos");
     }
@@ -125,7 +127,7 @@ REMORA::lin_eos (const Box& bx,
     });
 
     // Compute Brunt-Vaisala frequency (1/s2)
-    Real gorho0 = solverChoice.g / solverChoice.rho0;
+    Real gorho0 = g / solverChoice.rho0;
     // Really want enclosed nodes or something similar
     Box box_w = bx;
     box_w.surroundingNodes(2);
@@ -160,6 +162,8 @@ REMORA::nonlin_eos (const Box& bx,
                 const Array4<Real      >& rhoA,
                 const Array4<Real      >& rhoS,
                 const Array4<Real      >& bvf,
+                const Array4<Real      >& alpha,
+                const Array4<Real      >& beta,
                 const Array4<Real const>& Hz,
                 const Array4<Real const>& z_w,
                 const Array4<Real const>& z_r,
@@ -176,6 +180,12 @@ REMORA::nonlin_eos (const Box& bx,
     FArrayBox fab_bulk0(bx,1,amrex::The_Async_Arena()); auto bulk0 = fab_bulk0.array();
     FArrayBox fab_bulk1(bx,1,amrex::The_Async_Arena()); auto bulk1 = fab_bulk1.array();
     FArrayBox fab_bulk2(bx,1,amrex::The_Async_Arena()); auto bulk2 = fab_bulk2.array();
+    FArrayBox fab_DbulkDT(bx,1,amrex::The_Async_Arena()); auto DbulkDT = fab_DbulkDT.array();
+    FArrayBox fab_Dden1DT(bx,1,amrex::The_Async_Arena()); auto Dden1DT = fab_Dden1DT.array();
+    FArrayBox fab_DbulkDS(bx,1,amrex::The_Async_Arena()); auto DbulkDS = fab_DbulkDS.array();
+    FArrayBox fab_Dden1DS(bx,1,amrex::The_Async_Arena()); auto Dden1DS = fab_Dden1DS.array();
+
+    bool bulk_fluxes = solverChoice.bulk_fluxes;
 //
 //=======================================================================
 //  Non-linear equation of state.
[…]
