### 2025-03-04 bcbe12a
add bulk fluxes options to docs
