### 2025-03-05 291c884
add docs about bulk fluxes
