### 2025-03-05 d83f86e
add test for bulk fluxes
