+++ Exec/CMakeLists.txt
@@ -9,3 +9,4 @@ add_subdirectory(Upwelling)
 add_subdirectory(Channel_Test)
 add_subdirectory(DoubleGyre)
 add_subdirectory(IdealMiniGrid)
+add_subdirectory(BoundaryLayer)
