### 2025-03-06 6d48e22
make bulk fluxes work on GPU by making helper function static inline
