### 2025-03-07 0ec6d92
make a ProblemBase class and define virtual functions to be overriden by each problem where necessary
