+++ Exec/BoundaryLayer/prob.cpp
@@ -1,5 +1,4 @@
 #include "prob.H"
-#include "REMORA_prob_common.H"
 
 #include "AMReX_ParmParse.H"
 #include "AMReX_MultiFab.H"
@@ -8,21 +7,23 @@
 
 using namespace amrex;
 
-void
-amrex_probinit(
-  const amrex_real* /*problo*/,
-  const amrex_real* /*probhi*/)
+std::unique_ptr<ProblemBase>
+amrex_probinit(const amrex_real* problo, const amrex_real* probhi)
 {
+    return std::make_unique<Problem>(problo, probhi);
 }
+
+Problem::Problem(const amrex::Real* /*problo*/, const amrex::Real* /*probhi*/)
+{}
 
 /**
  * \brief Initializes bathymetry h and surface height Zeta
  */
-void
-init_custom_bathymetry (int lev, const Geometry& geom,
-                        MultiFab& mf_h, REMORA const& remora,
-                        const SolverChoice& m_solverChoice,
-                        int /*rrx*/, int /*rry*/)
+void Problem::init_analytic_bathymetry (
+        int lev, const amrex::Geometry& geom,
+        SolverChoice const& m_solverChoice,
+        REMORA const& remora,
+        amrex::MultiFab& mf_h)
 {
     auto geomdata = geom.data();
     bool EWPeriodic = geomdata.isPeriodic(0);
@@ -61,9 +62,11 @@ init_custom_bathymetry (int lev, const Geometry& geom,
     } // mfi
 }
 
-void
-init_custom_grid_scale (int /*lev*/, const Geometry& geom,
-                   MultiFab& mf_pm, MultiFab& mf_pn)
+void Problem::init_analytic_grid_scale (
+        int /*lev*/, const amrex::Geometry& geom,
+        SolverChoice const& m_solverChoice,
+        REMORA const& remora,
+        amrex::MultiFab& mf_pm, amrex::MultiFab& mf_pn)
 {
     auto geomdata = geom.data();
     const int Lm = geomdata.Domain().size()[0];
@@ -90,90 +93,94 @@ init_custom_grid_scale (int /*lev*/, const Geometry& geom,
 }
 
 /**
- * \brief Initializes coriolis factor
- */
-void
-init_custom_coriolis    (const Geometry& /*geom*/,
-                         MultiFab& /*mf_fcor*/,
-                         const SolverChoice& /*m_solverChoice*/) {}
-
-/**
  * \brief Initializes custom sea surface height
  */
-void
-init_custom_zeta (const Geometry& geom,
-                      MultiFab& mf_zeta,
-                      const SolverChoice& m_solverChoice)
+void Problem::init_analytic_zeta (
+        int /*lev*/, const amrex::Geometry& /*geom*/,
+        SolverChoice const& /*m_solverChoice*/,
+        REMORA const& /*remora*/,
+        MultiFab& mf_zeta)
 {
     mf_zeta.setVal(0.0_rt);
 }
 
-void
-init_custom_prob(
-        const Box& bx,
-        Array4<Real      > const& state,
-        Array4<Real      > const& x_vel,
-        Array4<Real      > const& y_vel,
-        Array4<Real      > const& z_vel,
-        Array4<Real const> const& /*z_w*/,
-        Array4<Real const> const& z_r,
-        Array4<Real const> const& /*Hz*/,
[…]
