+++ Exec/DoubleGyre/prob.cpp
@@ -8,131 +8,133 @@
 
 using namespace amrex;
 
-void
-amrex_probinit(
-  const amrex_real* /*problo*/,
-  const amrex_real* /*probhi*/)
+ProbParm parms;
+
+std::unique_ptr<ProblemBase>
+amrex_probinit(const amrex_real* problo, const amrex_real* probhi)
 {
+    return std::make_unique<Problem>(problo, probhi);
 }
+
+Problem::Problem(const amrex::Real* /*problo*/, const amrex::Real* /*probhi*/)
+{}
 
 /**
  * \brief Initializes bathymetry h and surface height Zeta
  */
-void
-init_custom_bathymetry (int /*lev*/, const Geometry& /*geom*/,
-                        MultiFab& mf_h, REMORA const& /*remora*/,
-                        const SolverChoice& /*m_solverChoice*/,
-                        int /*rrx*/, int /*rry*/)
+void Problem::init_analytic_bathymetry (
+        int /*lev*/, const amrex::Geometry& /*geom*/,
+        SolverChoice const& /*m_solverChoice*/,
+        REMORA const& /*remora*/,
+        amrex::MultiFab& mf_h)
 {
     mf_h.setVal(500.0_rt);
 }
 
-void
-init_custom_grid_scale (int /*lev*/, const Geometry& /*geom*/,
-                   MultiFab& /*mf_pm*/, MultiFab& /*mf_pn*/) {}
-
-/**
- * \brief Initializes coriolis factor
- */
-void
-init_custom_coriolis    (const Geometry& /*geom*/,
-                         MultiFab& /*mf_fcor*/,
-                         const SolverChoice& /*m_solverChoice*/) {}
-
 /**
  * \brief Initializes custom sea surface height
  */
-void
-init_custom_zeta (const Geometry& geom,
-                      MultiFab& mf_zeta,
-                      const SolverChoice& m_solverChoice)
+void Problem::init_analytic_zeta (
+        int /*lev*/, const amrex::Geometry& /*geom*/,
+        SolverChoice const& /*m_solverChoice*/,
+        REMORA const& /*remora*/,
+        MultiFab& mf_zeta)
 {
     mf_zeta.setVal(0.0_rt);
 }
 
-void
-init_custom_prob(
-        const Box& bx,
-        Array4<Real      > const& state,
-        Array4<Real      > const& x_vel,
-        Array4<Real      > const& y_vel,
-        Array4<Real      > const& z_vel,
-        Array4<Real const> const& /*z_w*/,
-        Array4<Real const> const& z_r,
-        Array4<Real const> const& /*Hz*/,
-        Array4<Real const> const& /*h*/,
-        Array4<Real const> const& /*Zt_avg1*/,
-        GeometryData const& geomdata,
-        const SolverChoice& m_solverChoice)
+void Problem::init_analytic_prob(
+        int lev,
+        const amrex::Geometry& geom,
+        SolverChoice const& m_solverChoice,
+        REMORA const& remora,
+        amrex::MultiFab& mf_cons,
+        amrex::MultiFab& mf_xvel,
+        amrex::MultiFab& mf_yvel,
+        amrex::MultiFab& mf_zvel)
 {
     bool l_use_salt = m_solverChoice.use_salt;
 
+    auto geomdata = geom.data();
     const int khi = geomdata.Domain().bigEnd()[2];
-
[…]
