### 2025-03-07 708b429
make rename custom to analytic and real to netcdf, but don't deprecate options yet
