+++ Source/TimeIntegration/REMORA_TimeStepML.cpp
@@ -170,7 +170,7 @@ REMORA::timeStepML (Real time, int /*iteration*/)
 
     scale_rhs_vars_inv();
 
-    if (solverChoice.coupling_type == CouplingType::TwoWay) {
+    if (solverChoice.coupling_type == CouplingType::two_way) {
         for (int lev=0; lev <= finest_level-1; lev++) {
             AverageDownTo(lev); // average lev+1 down to lev
         }
