+++ Source/REMORA.cpp
@@ -221,7 +221,7 @@ REMORA::post_timestep (int nstep, Real time, Real dt_lev0)
     particleData.Redistribute();
 #endif
 
-    if (solverChoice.coupling_type == CouplingType::TwoWay)
+    if (solverChoice.coupling_type == CouplingType::two_way)
     {
         for (int lev = finest_level-1; lev >= 0; lev--)
         {
@@ -260,7 +260,7 @@ REMORA::InitData ()
 
         InitFromScratch(start_time);
 
-        if (solverChoice.coupling_type == CouplingType::TwoWay) {
+        if (solverChoice.coupling_type == CouplingType::two_way) {
             AverageDown();
         }
 
@@ -271,7 +271,7 @@ REMORA::InitData ()
     }
 
     // Initialize flux registers (whether we start from scratch or restart)
-    if (solverChoice.coupling_type == CouplingType::TwoWay) {
+    if (solverChoice.coupling_type == CouplingType::two_way) {
         advflux_reg[0] = nullptr;
         for (int lev = 1; lev <= finest_level; lev++)
         {
@@ -444,11 +444,11 @@ REMORA::restart ()
 void
 REMORA::set_zeta (int lev)
 {
-    if (solverChoice.ic_bc_type == IC_BC_Type::Custom) {
+    if (solverChoice.ic_bc_type == IC_BC_Type::analytic) {
         prob->init_analytic_zeta(lev, geom[lev], solverChoice, *this, *vec_zeta[lev]);
 
 #ifdef REMORA_USE_NETCDF
-    } else if (solverChoice.ic_bc_type == IC_BC_Type::Real) {
+    } else if (solverChoice.ic_bc_type == IC_BC_Type::netcdf) {
         amrex::Print() << "Calling init_zeta_from_netcdf " << std::endl;
         init_zeta_from_netcdf(lev);
         amrex::Print() << "Sea surface height loaded from netcdf file \n " << std::endl;
@@ -466,11 +466,11 @@ REMORA::set_bathymetry (int lev)
     // Only set bathymetry on level 0, and interpolate for finer levels
     if (solverChoice.init_l0int_h) {
         if (lev==0) {
-            if (solverChoice.ic_bc_type == IC_BC_Type::Custom) {
+            if (solverChoice.ic_bc_type == IC_BC_Type::analytic) {
                 prob->init_analytic_bathymetry(lev, geom[lev], solverChoice, *this, *vec_hOfTheConfusingName[lev]);
 
 #ifdef REMORA_USE_NETCDF
-            } else if (solverChoice.ic_bc_type == IC_BC_Type::Real) {
+            } else if (solverChoice.ic_bc_type == IC_BC_Type::netcdf) {
                 amrex::Print() << "Calling init_bathymetry_from_netcdf " << std::endl;
                 init_bathymetry_from_netcdf(lev);
                 amrex::Print() << "Bathymetry loaded from netcdf file \n " << std::endl;
@@ -487,11 +487,11 @@ REMORA::set_bathymetry (int lev)
             FillCoarsePatch(lev,dummy_time,vec_hOfTheConfusingName[lev].get(), vec_hOfTheConfusingName[lev-1].get());
         }
     } else if (solverChoice.init_ana_h) {
-        if (solverChoice.ic_bc_type == IC_BC_Type::Custom) {
+        if (solverChoice.ic_bc_type == IC_BC_Type::analytic) {
             prob->init_analytic_bathymetry(lev, geom[lev], solverChoice, *this,*vec_hOfTheConfusingName[lev]);
 
 #ifdef REMORA_USE_NETCDF
[…]
