+++ Source/REMORA_DataStruct.H
@@ -12,7 +12,7 @@
 #include "REMORA_IndexDefines.H"
 
 enum struct CouplingType {
-    OneWay, TwoWay
+    one_way, two_way
 };
 
 enum class Coord {
@@ -28,11 +28,11 @@ enum class AdvectionScheme {
 };
 
 enum class IC_BC_Type {
-    Custom, Real
+    analytic, netcdf
 };
 
 enum class Cor_Type {
-    Custom, Beta_Plane, Real
+    analytic, beta_plane, netcdf
 };
 
 enum class PlotfileType {
@@ -40,11 +40,11 @@ enum class PlotfileType {
 };
 
 enum class VertMixingType {
-    analytical, GLS
+    analytic, GLS
 };
 
 enum class HorizMixingType {
-    analytical, constant
+    analytic, constant
 };
 
 enum class GLS_StabilityType {
@@ -60,7 +60,7 @@ enum class BottomStressType {
 };
 
 enum class GridScaleType {
-    Constant, Custom
+    constant, analytic
 };
 
 struct SolverChoice {
@@ -183,22 +183,29 @@ struct SolverChoice {
         pp.query("grid_scale_type",grid_scale_type_string);
 
         if (amrex::toLower(grid_scale_type_string) == "constant") {
-            grid_scale_type = GridScaleType::Constant;
+            grid_scale_type = GridScaleType::constant;
         } else if (amrex::toLower(grid_scale_type_string) == "custom") {
-            grid_scale_type = GridScaleType::Custom;
+            amrex::Warning("Initialization of grid scale from prob.cpp is now called 'analytic'. 'custom' will be deprecated");
+            grid_scale_type = GridScaleType::analytic;
+        } else if (amrex::toLower(grid_scale_type_string) == "analytic") {
+            grid_scale_type = GridScaleType::analytic;
         } else {
             amrex::Error("Don't know this grid_scale_type");
         }
 
-        static std::string ic_bc_type_string = "custom";
+        static std::string ic_bc_type_string = "analytic";
         pp.query("ic_bc_type", ic_bc_type_string);
 
-        if ( !ic_bc_type_string.compare("Custom") ||
-             !ic_bc_type_string.compare("custom") ) {
-            ic_bc_type = IC_BC_Type::Custom;
-        } else if ( !ic_bc_type_string.compare("Real") ||
-                    !ic_bc_type_string.compare("real") ) {
-            ic_bc_type = IC_BC_Type::Real;
+        if ( amrex::toLower(ic_bc_type_string) == "custom") {
+            amrex::Warning("Problem initialization from prob.cpp is now called 'analytic'. 'custom' will be deprecated");
+            ic_bc_type = IC_BC_Type::analytic;
+        } else if ( amrex::toLower(ic_bc_type_string) == "analytic") {
+            ic_bc_type = IC_BC_Type::analytic;
+        } else if ( amrex::toLower(ic_bc_type_string) == "netcdf") {
+            ic_bc_type = IC_BC_Type::netcdf;
+        } else if ( amrex::toLower(ic_bc_type_string) == "real") {
+            amrex::Warning("Problem initialization from NetCDF (remora.ic_bc_type) is now called 'netcdf'. 'real' will be deprecated");
+            ic_bc_type = IC_BC_Type::netcdf;
         } else {
             amrex::Error("Don't know this ic_bc_type");
         }
@@ -206,10 +213,12 @@ struct SolverChoice {
         // Which type of refinement
         static std::string coupling_type_string = "TwoWay";
         pp.query("coupling_type",coupling_type_string);
-        if (coupling_type_string == "TwoWay") {
-            coupling_type = CouplingType::TwoWay;
[…]
