+++ Source/BoundaryConditions/REMORA_FillPatch.cpp
@@ -126,7 +126,7 @@ REMORA::FillPatch (int lev, Real time, MultiFab& mf_to_fill, Vector<MultiFab*> c
 
 #ifdef REMORA_USE_NETCDF
         // Fill the data which is stored in the boundary data read from netcdf files
-        if ( (solverChoice.ic_bc_type == IC_BC_Type::Real) && (lev==0) &&
+        if ( (solverChoice.ic_bc_type == IC_BC_Type::netcdf) && (lev==0) &&
              (bdy_var_type != BdyVars::null) )
         {
             fill_from_bdyfiles(mf_to_fill,*mask,time,bccomp,bdy_var_type, icomp,icomp_calc,mf_calc,dt);

[…]
