### 2025-03-08 bb380ab
added blankproblem and added to docs about analytic functions
