+++ Exec/BlankProblem/CMakeLists.txt
@@ -1,0 +1,13 @@
+set(remora_exe_name blankproblem)
+
+add_executable(${remora_exe_name} "")
+target_sources(${remora_exe_name}
+   PRIVATE
+     prob.H
+     prob.cpp
+)
+
+target_include_directories(${remora_exe_name} PRIVATE ${CMAKE_CURRENT_SOURCE_DIR})
+
+include(${CMAKE_SOURCE_DIR}/CMake/BuildREMORAExe.cmake)
+build_remora_exe(${remora_exe_name})
