### 2025-03-17 d1212be
adjust buoy2 condition for BVF stability
