### 2025-03-24 #7737 @jatkinson1000
Hi @hklion ,

Thanks for your submission to JOSS, I will be your editor.

I have taken a look at the repository and confirm it:

- [x] has a general description
- [x] is within scope
- [ ] has an OSI-approved open-source license
- [x] is of sufficient scholarly effort
- [x] has a paper of appropriate length for JOSS

I note that the license isn't a default OSI one, as far as I can tell, but seems to be a variation on [LANL BSD-3](https://opensource.org/license/los-alamos-national-labs-bsd-3-variant). I checked this with @kthyng and they said this is probably OK, but if you could clarify this (perhaps it is an older version?) that would be useful, thanks.

In the meantime I'll look to appoint some reviewers.
