### 2025-03-24 #7737 @hklion
Hi @jatkinson1000,

The license is the [LBNL BSD-3](https://opensource.org/license/bsd-3-clause-lbnl), so very similar to the LANL one, but its own thing.

Thanks!
