### 2025-03-25 #7737 @bastorer
Hello all,

This is a touch outside my wheelhouse (mostly the AMR and CUDA), but it looks pretty interesting and I'd be happy to be a reviewer.
