### 2025-03-26 9bc60e2
Add framework to read in time series data from netcdf file
- NCTimeSeries reads in two time points and linearly interpolates
- Adds option to read in winds from netcdf file
