### 2025-03-26 153afb7
add needed compile guards
