+++ Source/IO/REMORA_NCTimeSeries.cpp
@@ -1,0 +1,149 @@
+#include "REMORA_NCTimeSeries.H"
+#include "REMORA_NCFile.H"
+
+#include "AMReX_ParallelDescriptor.H"
+
+#include <string>
+
+#ifdef REMORA_USE_NETCDF
+NCTimeSeries::NCTimeSeries (const std::string a_file_name, const std::string a_field_name,
+                            const amrex::Box& a_domain,
+                            amrex::MultiFab* a_mf_var, bool a_is2d, bool a_save_interpolated) {
+    file_name = a_file_name;
+    field_name = a_field_name;
+    domain = a_domain;
+    mf_var = a_mf_var;
+    is2d = a_is2d;
+    save_interpolated = a_save_interpolated;
+}
+
+void NCTimeSeries::Initialize() {
+    // open file
+    amrex::Print() << "Loading " << field_name << " from NetCDF file " << file_name << std::endl;
+
+    // Check units of time stamps; should be days
+    std::string unit_str = ReadNetCDFVarAttrStr(file_name, "ocean_time", "units"); // works on proc 0
+    if (amrex::ParallelDescriptor::IOProcessor())
+    {
+        if (unit_str.find("days") == std::string::npos) {
+            amrex::Print() << "Units of ocean_time given as: " << unit_str << std::endl;
+            amrex::Abort("Units must be in days.");
+        }
+    }
+    // get times and put in array
+    using RARRAY = NDArray<amrex::Real>;
+    amrex::Vector<RARRAY> array_ts(1);
+    ReadNetCDFFile(file_name, {"ocean_time"}, array_ts); // filled only on proc 0
+    if (amrex::ParallelDescriptor::IOProcessor())
+    {
+        int ntimes_io = array_ts[0].get_vshape()[0];
+        for (int nt(0); nt < ntimes_io; nt++)
+        {
+            // Convert ocean time from days to seconds
+            ocean_times.push_back((*(array_ts[0].get_data() + nt)) * amrex::Real(60.0) * amrex::Real(60.0) * amrex::Real(24.0));
+            // amrex::Print() << "TIMES " << ocean_times[nt] << std::endl;
+        }
+    }
+    int ntimes = ocean_times.size();
+    int ioproc = amrex::ParallelDescriptor::IOProcessorNumber();
+    amrex::ParallelDescriptor::Bcast(&ntimes,1,ioproc);
+    if (!(amrex::ParallelDescriptor::IOProcessor())) {
+        ocean_times.resize(ntimes);
+    }
+    amrex::ParallelDescriptor::Bcast(ocean_times.data(), ocean_times.size(), ioproc);
+
+    // Initialize MultiFabs
+    // This assumes that the level 0 distribution map and box array won't
+    mf_before = new amrex::MultiFab(mf_var->boxArray(), mf_var->DistributionMap(), 1, mf_var->nGrowVect());
+    mf_after = new amrex::MultiFab(mf_var->boxArray(), mf_var->DistributionMap(), 1, mf_var->nGrowVect());
+    if (save_interpolated) {
+        mf_interpolated = new amrex::MultiFab(mf_var->boxArray(), mf_var->DistributionMap(), 1, mf_var->nGrowVect());
+    }
+
+    // dummy initialization
+    i_time_before = -100;
+}
+
+void NCTimeSeries::update_interpolated_to_time (amrex::Real time) {
+    // Figure out time index:
+    AMREX_ASSERT(time >= ocean_times[0]);
+    AMREX_ASSERT(time <= ocean_times[ocean_times.size()-1]);
+    int i_time_before_old = i_time_before;
[…]
