### 2025-03-27 #7737 @jatkinson1000
Hi there @esclapez!

This paper has been submitted to JOSS, the Journal of Open Source Software.
It is a code based off-of the ROMS ocean model, but specifically designed to target GPU architectures using the AMReX meshing framework.

Would you be willing to review it?
Please indicate by replying to this thread.

You can find information about reviewing for JOSS [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
