### 2025-03-28 faabbac
add docs about time series winds and adjust time series name
