### 2025-03-30 #7737 @esclapez
@jatkinson1000 Thank you for the invite. I can review this submission. I've worked with AMReX in the past and on ocean modeling more recently.
