### 2025-03-31 #7958 @jatkinson1000
Hi @hklion @jwallwork23 @bastorer @esclapez this is the review thread for the paper.
All of our communications will happen here from now on.

```@editorialbot generate my checklist```

These checklists contain the JOSS requirements.
As you go over the submission, please check off anything that you feel has been satisfied for publication in JOSS. The first comment in this thread also contains links to the [JOSS reviewer guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).

The JOSS review is different from some other journals. Our goal is to work with the authors to help them meet our criteria instead of merely passing judgment on the submission. As such, the reviewers are encouraged to submit issues and pull requests on the software repository. When doing so, please mention `openjournals/joss-reviews#7958` so that a link is created to this thread. Please also feel free to comment and ask questions on this thread. You are welcome to post comments/questions/suggestions as you come across them instead of waiting until you've reviewed the entire package.

We aim for reviews to be completed within about 2-4 weeks. Please let me know if any of you require some more time.
We can also use EditorialBot (our bot) to set automatic reminders if you know you'll be away for a known period of time.
I'll check back in in a fortnight to see how things are progressing.
