### 2025-04-14 #7958 @jatkinson1000
Hi @hklion @jwallwork23 @bastorer @esclapez, just checking in to see how this is going?

@bastorer please could you generate your checklist using `@editorialbot generate my checklist`?

We'd ideally aim for a 4-week turnaround on reviews which would take us to the 28th April, but please let me know if you expect to take longer.
