### 2025-04-14 a8a887a
Add nudging to climatology
