### 2025-04-16 069497c
add climatology documentation
