+++ Docs/sphinx_doc/Inputs.rst
@@ -83,31 +83,31 @@ Currently, if one initial, grid, or boundary files are specified, they all must 
 List of Parameters
 ------------------
 
-+---------------------------+-------------------------------+-------------+---------------------------+
-| Parameter                 | Definition                    | Acceptable  | Default                   |
-|                           |                               | Values      |                           |
-+===========================+===============================+=============+===========================+
-| **remora.ic_bc_type**     | read initial, grid, and       |             |                           |
-|                           | boundary data from NetCDF     | true/false  | false                     |
-|                           | files                         |             |                           |
-+---------------------------+-------------------------------+-------------+---------------------------+
-| **remora.nc_init_file_0** | initial data NetCDF file name | string      | must be set               |
-|                           |                               |             | if ``remora.ic_bc_type``  |
-|                           |                               |             | is true                   |
-+---------------------------+-------------------------------+-------------+---------------------------+
-| **remora.nc_grid_file_0** | grid data NetCDF file name    | string      | must be set               |
-|                           |                               |             | if ``remora.ic_bc_type``  |
-|                           |                               |             | is true                   |
-+---------------------------+-------------------------------+-------------+---------------------------+
-| **remora.nc_bdry_file_0** | boundary data NetCDF file     | string      | must be set               |
-|                           | name                          |             | if ``remora.ic_bc_type``  |
-|                           |                               |             | is true                   |
-+---------------------------+-------------------------------+-------------+---------------------------+
-| **remora.nc_frc_file**    | forcing data NetCDF file name | string      | must be set               |
-|                           |                               |             | if ``remora.wind_type``   |
-|                           |                               |             | or ``remora.smflux_type`` |
-|                           |                               |             | equal ``netcdf``          |
-+---------------------------+-------------------------------+-------------+---------------------------+
++-------------------------------+-----------------------------------+-------------+---------------------------+
+| Parameter                     | Definition                        | Acceptable  | Default                   |
+|                               |                                   | Values      |                           |
++===============================+===================================+=============+===========================+
+| **remora.ic_bc_type**         | read initial, grid, and           |             |                           |
+|                               | boundary data from NetCDF         | true/false  | false                     |
+|                               | files                             |             |                           |
++-------------------------------+-----------------------------------+-------------+---------------------------+
+| **remora.nc_init_file_0**     | initial data NetCDF file name     | string      | must be set               |
+|                               |                                   |             | if ``remora.ic_bc_type``  |
+|                               |                                   |             | is true                   |
++-------------------------------+-----------------------------------+-------------+---------------------------+
+| **remora.nc_grid_file_0**     | grid data NetCDF file name        | string      | must be set               |
+|                               |                                   |             | if ``remora.ic_bc_type``  |
+|                               |                                   |             | is true                   |
++-------------------------------+-----------------------------------+-------------+---------------------------+
+| **remora.nc_bdry_file_0**     | boundary data NetCDF file         | string      | must be set               |
+|                               | name                              |             | if ``remora.ic_bc_type``  |
+|                               |                                   |             | is true                   |
++-------------------------------+-----------------------------------+-------------+---------------------------+
+| **remora.nc_frc_file**        | forcing data NetCDF file name     | string      | must be set               |
+|                               |                                   |             | if ``remora.wind_type``   |
+|                               |                                   |             | or ``remora.smflux_type`` |
+|                               |                                   |             | equal ``netcdf``          |
++-------------------------------+-----------------------------------+-------------+---------------------------+
 
 Resolution and Tiling
 =====================
@@ -801,6 +801,43 @@ List of Parameters
 
 These parameters are used to calculate the vertical S-grid stretch/transform functions detailed in
 :ref:`Vertical S-Coordinate<VerticalSCoord>`.
+
+Climatology parameters
[…]
