### 2025-04-16 22890e9
add info about climatology and wind regtests
