+++ Docs/sphinx_doc/RegressionTests.rst
@@ -216,191 +216,247 @@ Based on :ref:`Doubly Periodic<doublyperiodic>`, which always includes Coriolis:
 
 Based on :ref:`Ideal Mini Grid<idealminigrid>`, which always includes Coriolis and PnetCDF:
 
-+----------------------------------------+--------------+------------------+-------------------+----------------------------------+
-| Test                                   | nx ny nz     | xbc              | ybc               | Other                            |
-+========================================+==============+==================+===================+==================================+
-| IdealMiniGrid                          | 10 16 20     | Clamped          | Clamped           | MPI                              |
-|                                        |              |                  |                   |                                  |
-|                                        |              |                  |                   | Varying salt at boundary         |
-+----------------------------------------+--------------+------------------+-------------------+----------------------------------+
-| IdealMiniGrid-1grid                    | 10 16 20     | Clamped          | Clamped           | Varying salt at boundary         |
-+----------------------------------------+--------------+------------------+-------------------+----------------------------------+
-| IdealMiniGrid-CF-Uvel-OMP              | 10 16 20     | Chapman-Flather  | Chapman-Flather   | MPI + OpenMP                     |
-|                                        |              |                  |                   |                                  |
-|                                        |              |                  |                   | Varying velocity at boundary     |
-+----------------------------------------+--------------+------------------+-------------------+----------------------------------+
-| IdealMiniGrid-CFO-Salt-OMP             | 10 16 20     | Chapman-Flather  | Chapman-Flather   | MPI + OpenMP                     |
-|                                        |              |                  |                   |                                  |
-|                                        |              | Radiation        | Radiation         | Varying salt at boundary         |
-+----------------------------------------+--------------+------------------+-------------------+----------------------------------+
-| IdealMiniGrid-CFO-Temp-OMP             | 10 16 20     | Chapman-Flather  | Chapman-Flather   | MPI + OpenMP                     |
-|                                        |              |                  |                   |                                  |
-|                                        |              | Radiation        | Radiation         | Varying temperature at boundary  |
-+----------------------------------------+--------------+------------------+-------------------+----------------------------------+
-| IdealMiniGrid-CFO-Uvel-OMP             | 10 16 20     | Chapman-Flather  | Chapman-Flather   | MPI + OpenMP                     |
-|                                        |              |                  |                   |                                  |
-|                                        |              | Radiation        | Radiation         | Varying velocity at boundary     |
-+----------------------------------------+--------------+------------------+-------------------+----------------------------------+
-| IdealMiniGrid-EWWall-OMP               | 10 16 20     | SlipWall         | Clamped           | MPI + OpenMP                     |
-|                                        |              |                  |                   |                                  |
-|                                        |              |                  |                   | Varying salt at boundary         |
-+----------------------------------------+--------------+------------------+-------------------+----------------------------------+
-| IdealMiniGrid-EWWall-restart           | 10 16 20     | SlipWall         | Clamped           | MPI                              |
-|                                        |              |                  |                   |                                  |
-|                                        |              |                  |                   | Varying salt at boundary         |
-|                                        |              |                  |                   |                                  |
-|                                        |              |                  |                   | restart                          |
-+----------------------------------------+--------------+------------------+-------------------+----------------------------------+
-| IdealMiniGrid-NSWall-OMP               | 10 16 20     | Clamped          | SlipWall          | MPI + OpenMP                     |
-|                                        |              |                  |                   |                                  |
-|                                        |              |                  |                   | Varying salt at boundary         |
-+----------------------------------------+--------------+------------------+-------------------+----------------------------------+
-| IdealMiniGrid-NSWall-restart           | 10 16 20     | Clamped          | Slipwall          | MPI                              |
-|                                        |              |                  |                   |                                  |
[…]
