### 2025-04-17 df65e77
get build system ready to include fortran
