+++ .github/workflows/hip.yml
@@ -57,6 +57,7 @@ jobs:
         hipcc --version
         which clang
         which clang++
+        which flang
 
         cmake \
           -B${{runner.workspace}}/REMORA/build-${{matrix.os}} \
@@ -74,6 +75,7 @@ jobs:
           -DCODECOV:BOOL=ON \
           -DCMAKE_C_COMPILER=$(which clang)             \
           -DCMAKE_CXX_COMPILER=$(which clang++)         \
+          -DCMAKE_Fortran_COMPILER=$(which flang)         \
           -DCMAKE_CXX_STANDARD=17 \
           ${{github.workspace}};
         # ${{matrix.mpipreflags}} \
