+++ CMake/SetAmrexOptions.cmake
@@ -26,7 +26,11 @@ set(AMReX_PROFPARSER OFF)
 set(AMReX_CUDA ${REMORA_ENABLE_CUDA})
 set(AMReX_ACC OFF)
 set(AMReX_PLOTFILE_TOOLS ${REMORA_ENABLE_FCOMPARE})
+
 set(AMReX_FORTRAN OFF)
+if(REMORA_ENABLE_FUNWAVE_FORT) 
+  set(AMReX_FORTRAN ON)
+endif()
 
 set(AMReX_PARTICLES OFF)
 if(REMORA_ENABLE_PARTICLES)
