### 2025-04-18 f4d8b4d
removed mask_psi loading
