+++ Source/Initialization/REMORA_init_from_netcdf.cpp
@@ -20,7 +20,7 @@ read_data_from_netcdf (int /*lev*/, const Box& domain, const std::string& fname,
 void
 read_masks_from_netcdf (int /*lev*/, const Box& domain, const std::string& fname,
                        FArrayBox& NC_mskr_fab, FArrayBox& NC_msku_fab,
-                       FArrayBox& NC_mskv_fab, FArrayBox& NC_mskp_fab);
+                       FArrayBox& NC_mskv_fab);
 
 Real
 read_bdry_from_netcdf (const Box& domain, const std::string& fname,
@@ -386,13 +386,12 @@ REMORA::init_masks_from_netcdf (int lev)
     Vector<FArrayBox> NC_mskr_fab     ; NC_mskr_fab.resize(num_boxes_at_level[lev]);
     Vector<FArrayBox> NC_msku_fab     ; NC_msku_fab.resize(num_boxes_at_level[lev]);
     Vector<FArrayBox> NC_mskv_fab     ; NC_mskv_fab.resize(num_boxes_at_level[lev]);
-    Vector<FArrayBox> NC_mskp_fab     ; NC_mskp_fab.resize(num_boxes_at_level[lev]);
 
     for (int idx = 0; idx < num_boxes_at_level[lev]; idx++)
     {
         read_masks_from_netcdf(lev,boxes_at_level[lev][idx], nc_grid_file[lev][idx],
                                     NC_mskr_fab[idx],NC_msku_fab[idx],
-                                    NC_mskv_fab[idx],NC_mskp_fab[idx]);
+                                    NC_mskv_fab[idx]);
 
 #ifdef _OPENMP
 #pragma omp parallel if (amrex::Gpu::notInLaunchRegion())
@@ -404,7 +403,6 @@ REMORA::init_masks_from_netcdf (int lev)
             FArrayBox &mskr_fab  = (*vec_mskr[lev])[mfi];
             FArrayBox &msku_fab  = (*vec_msku[lev])[mfi];
             FArrayBox &mskv_fab  = (*vec_mskv[lev])[mfi];
-            FArrayBox &mskp_fab  = (*vec_mskp[lev])[mfi];
 
             //
             // FArrayBox to FArrayBox copy does "copy on intersection"
