### 2025-04-21 3bfc65d
allow user to specify time variable name in netcdf file
