### 2025-04-24 #7958 @jwallwork23
Hi @hklion, thanks for this submission. I'm in the process of installing REMORA. In the meantime, I have a couple of requests based on reading through the paper, as follows. Thanks!

From the JOSS review checklist:
> **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

The statement of need mentions that *most* ocean modelling codes don't have the ability to use GPU acceleration. Could you give examples of any that do?

From the JOSS review checklist:
> **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

The features list is impressive. However, a long bullet proof list doesn't seem to be the best suited format for a paper. Might the authors consider reformatting this? Bullets could still be used, but groupings into (e.g.) timestepping, advection schemes, boundary conditions, etc. might make it more readable.
