### 2025-04-24 seahorce-scidac/ROMSX#375 @joewallwork
[JOSS Review]: Docstrings
Towards https://github.com/openjournals/joss-reviews/issues/7958.

It's great that `REMORA` includes extensive documentation, including API documentation that can be generated using Doxygen. However, I feel that the API documentation isn't as useful as it could be because the docstrings don't give any information about the input variables. For example, `REMORA::advance_2d` has a description for the `lev` input argument, but not any of the others, meaning it's unclear what these variables represent and what values should be passed.

It'd be great if variable descriptions could be added to the docstrings to improve the API documentation.
