### 2025-04-24 #7958 @jwallwork23
I've worked through the review checklist and am happy with everything, except for the items above, one item that I raised in https://github.com/seahorce-scidac/REMORA/issues/375, and also that I want to test the install and tests on GPU. (I'll aim to do this tomorrow.)
