### 2025-04-25 687d55b
only read in boundary data required by bounday config and fix floating point issue in dt cheeck
