+++ Docs/Doxyfile.in
@@ -44,7 +44,7 @@ PROJECT_NUMBER         =
 # for a project that appears at the top of each page and should give viewer a
 # quick idea about the purpose of the project. Keep the description short.
 
-PROJECT_BRIEF          = "Energy Research and Forecasting: An Atmospheric Modeling Code"
+PROJECT_BRIEF          = "Regional Modeling of Oceans Refined Adaptively"
 
 
 # With the PROJECT_LOGO tag one can specify a logo or an icon that is included
@@ -782,7 +782,7 @@ INPUT                  = main.dox \
 #                         ../../amrex/Src/AMRex_MultiFab.H \
 #                         ../../amrex/Src/AMReX_IntVect.H \
 #                         ../../amrex/Src/Base/AMReX_RealBox.H \
-#                         ../../amrex/Src/EB/AMReX_EB2_GeometryShop.H 
+#                         ../../amrex/Src/EB/AMReX_EB2_GeometryShop.H
 
 # This tag can be used to specify the character encoding of the source files
 # that doxygen parses. Internally doxygen uses the UTF-8 encoding. Doxygen uses
