+++ Source/REMORA.H
@@ -67,25 +67,20 @@ namespace InterpType {
     };
 }
 
-// Forward declaration to resolve circular dependency
+// Forward declarations to resolve circular dependency
 class ProblemBase;
-
 #ifdef REMORA_USE_NETCDF
 class NCTimeSeries;
 #endif
 
+/*!
+ * Class that stores all relevant simulation state data with methods for time stepping
+ */
 class REMORA
     : public amrex::AmrCore
 {
 public:
 
-    ////////////////
-    // public member functions
-
-    /**
-     * constructor - reads in parameters from inputs file
-     *             - sizes multilevel arrays and data structures
-     */
     REMORA ();
     virtual ~REMORA();
 
@@ -112,12 +107,12 @@ public:
     /** Integrate conserved quantities for diagnostics */
     void sum_integrated_quantities(amrex::Real time);
 
-    /* Perform the volume-weighted sum */
+    /** Perform the volume-weighted sum */
     amrex::Real
     volWgtSumMF(int lev,
       const amrex::MultiFab& mf, int comp, bool local, bool finemask);
 
-    /* Decide if it is time to take an action */
+    /** Decide if it is time to take an action */
     bool is_it_time_for_action(int nstep, amrex::Real time, amrex::Real dt,
                                int action_interval, amrex::Real action_per);
 
@@ -171,9 +166,10 @@ public:
     /** Make mask to zero out covered cells*/
     amrex::MultiFab& build_fine_mask(int lev);
 
-    /** write plotfile */
+    /** main driver for writing AMReX plotfiles */
     void WritePlotFile ();
 
+    /** write out particular data to an AMReX plotfile */
     void WriteMultiLevelPlotfileWithBathymetry (const std::string &plotfilename,
                                                 int nlevels,
                                                 const amrex::Vector<const amrex::MultiFab*> &mf,
@@ -187,6 +183,7 @@ public:
                                                 const amrex::Vector<std::string>& extra_dirs = amrex::Vector<std::string>()) const;
 
 
+    /** write out header data for an AMReX plotfile */
     void WriteGenericPlotfileHeaderWithBathymetry (std::ostream &HeaderFile,
                                                    int nlevels,
                                                    const amrex::Vector<amrex::BoxArray> &bArray,
@@ -197,16 +194,25 @@ public:
                                                    const std::string &levelPrefix,
                                                    const std::string &mfPrefix) const;
 
+    /** default prefix for input file parameters */
     std::string pp_prefix {"remora"};
 
+    /** multilevel data container for last step's scalar data: temperature, salinity, passive scalar  */
     amrex::Vector<amrex::MultiFab*> cons_old;
+    /** multilevel data container for last step's x velocities (u in ROMS) */
     amrex::Vector<amrex::MultiFab*> xvel_old;
+    /** multilevel data container for last step's y velocities (v in ROMS) */
     amrex::Vector<amrex::MultiFab*> yvel_old;
+    /** multilevel data container for last step's z velocities (largely unused; W stored separately) */
     amrex::Vector<amrex::MultiFab*> zvel_old;
 
[…]
