### 2025-04-29 #7958 @esclapez
@jatkinson1000 I am digging in REMORA now and should be done by the end of the week, sorry for the slight delay.

In disclosure of any COI:
 - one of the authors was my line manager some 5 years ago (I have changed institution and continent since then)
 - another author and myself are co-authors on a recent paper related to work dating back to 2022. The focus of that paper was on another software (SUNDIALS), and our respective contributions centered around how SUNDIALS was integrated in different applications code we were developing separately around that time.

Please let me know if you think those points prevent me from being able to review this submission.
