### 2025-04-30 5431f60
add docstrings to REMORA.H and do some cleanup of unused class members
