+++ Source/REMORA.H
@@ -82,7 +82,7 @@ class REMORA
 public:
 
     REMORA ();
-    virtual ~REMORA();
+    virtual ~REMORA ();
 
     /** Advance solution to final time */
     void Evolve ();
@@ -105,15 +105,15 @@ public:
     // Diagnostics
 
     /** Integrate conserved quantities for diagnostics */
-    void sum_integrated_quantities(amrex::Real time);
+    void sum_integrated_quantities (amrex::Real time);
 
     /** Perform the volume-weighted sum */
     amrex::Real
-    volWgtSumMF(int lev,
+    volWgtSumMF (int lev,
       const amrex::MultiFab& mf, int comp, bool local, bool finemask);
 
     /** Decide if it is time to take an action */
-    bool is_it_time_for_action(int nstep, amrex::Real time, amrex::Real dt,
+    bool is_it_time_for_action (int nstep, amrex::Real time, amrex::Real dt,
                                int action_interval, amrex::Real action_per);
 
     /** Make a new level using provided BoxArray and DistributionMapping and
@@ -163,7 +163,7 @@ public:
                         const amrex::Geometry fine_geom,
                         const amrex::Real dt, const amrex::Real time);
 
-    /** Make mask to zero out covered cells*/
+    /** Make mask to zero out covered cells (for mesh refinement) */
     amrex::MultiFab& build_fine_mask(int lev);
 
     /** main driver for writing AMReX plotfiles */
@@ -478,6 +478,7 @@ public:
                      const int N,
                      const amrex::Real dt_lev);
 
+    /** Calculate bulk temperature, salinity, wind fluxes */
     void bulk_fluxes (int lev,
                       amrex::MultiFab* mf_cons,
                       amrex::MultiFab* mf_uwind,
@@ -516,7 +517,7 @@ public:
                   const int nnew, int nstp, int nrhs,
                   int N, const amrex::Real dt_lev);
 
-    /** Prestep calculations for the tracers */
+    /** Prestep advection calculations for the tracers */
     void prestep_t_advection (const amrex::Box& tbx,
                         const amrex::Box& gbx,
                         const amrex::Array4<amrex::Real      >& tempold,
@@ -537,6 +538,7 @@ public:
                         int iic, int ntfirst, int nrhs, int N,
                         const amrex::Real dt_lev);
 
+    /** RHS terms for tracer */
     void rhs_t_3d (const amrex::Box& bx,
                    const amrex::Box& gbx,
                    const amrex::Array4<amrex::Real      >& t,
@@ -552,7 +554,7 @@ public:
                    const amrex::Array4<amrex::Real const>& mskv,
                    int nrhs, int nnew, int N, const amrex::Real dt_lev);
 
-    /** Calculation of the RHS */
+    /** RHS terms for 3D momentum */
     void rhs_uv_3d (const amrex::Box& xbx,
                     const amrex::Box& ybx,
                     const amrex::Array4<amrex::Real const>& uold,
@@ -573,6 +575,7 @@ public:
                     const amrex::Array4<amrex::Real      >& FC,
                     int nrhs, int N);
 
+    /** RHS terms for 2D momentum */
     void rhs_uv_2d (const amrex::Box& xbx,
                     const amrex::Box& ybx,
                     const amrex::Array4<amrex::Real const>& uold,
@@ -583,6 +586,7 @@ public:
                     const amrex::Array4<amrex::Real const>& Dvom,
                     const int nrhs);
 
+    /** Wrapper around equation of state calculation */
     void rho_eos (const amrex::Box& bx,
                   const amrex::Array4<amrex::Real const>& state,
                   const amrex::Array4<amrex::Real      >& rho,
@@ -598,6 +602,7 @@ public:
                   const amrex::Array4<amrex::Real const>& mskr,
[…]
