### 2025-05-09 #7958 @jwallwork23
> I've worked through the review checklist and am happy with everything, except for the items above, one item that I raised in [seahorce-scidac/REMORA#375](https://github.com/seahorce-scidac/REMORA/issues/375), and also that I want to test the install and tests on GPU. (I'll aim to do this tomorrow.)

Sorry, it took me longer than expected to get round to this. I built with GPU support on an Nvidia device yesterday and it seemed to go fine. I notice that there are [GPU regression tests](https://remora-ocean.readthedocs.io/en/latest/RegressionTests.html#nightly-regression-tests-on-gpu) that are regularly run. @hklion please could you advise on how to run these? I couldn't see any details for running GPU tests on that page or the [testing page](https://remora-ocean.readthedocs.io/en/latest/testing.html#testing).
