### 2025-05-09 0fe5018
add docs about GPU CTest
