### 2025-05-10 #7958 @hklion
@jwallwork23 Glad to hear your NVIDIA install went well! I just added instructions for how to run the ctest suite with GPUs to the testing page. I'd recommend those since they're the same tests run on github CI. I've run something similar on my local machine, but let me know if you run into any issues that I can help with.

Apologies for the delay in responding to your other comments. April/May suddenly became very busy. I'm currently working on implementing the recommended changes to the docstrings and paper, and hope to have it done by the end of next week.
