### 2025-05-12 812d233
remove unused variables and update docstrings
