+++ Source/Initialization/REMORA_make_new_level.cpp
@@ -9,10 +9,17 @@
 
 using namespace amrex;
 
-// Make a new level using provided BoxArray and DistributionMapping and
-// fill with interpolated coarse level data (overrides the pure virtual function in AmrCore)
-// regrid  --> RemakeLevel            (if level already existed)
-// regrid  --> MakeNewLevelFromCoarse (if adding new level)
+/**
+ * Make a new level using provided BoxArray and DistributionMapping and
+ * fill with interpolated coarse level data (overrides the pure virtual function in AmrCore)
+ * regrid  --> RemakeLevel            (if level already existed)
+ * regrid  --> MakeNewLevelFromCoarse (if adding new level)
+ *
+ * @param[in   ] lev     level to make
+ * @param[in   ] time    current time
+ * @param[in   ] ba      BoxArray for the level
+ * @param[in   ] dm      DistributionMapping for the level
+ */
 void
 REMORA::MakeNewLevelFromCoarse (int lev, Real time, const BoxArray& ba,
                                const DistributionMapping& dm)
@@ -96,9 +103,15 @@ REMORA::MakeNewLevelFromCoarse (int lev, Real time, const BoxArray& ba,
 #endif
 }
 
-// Remake an existing level using provided BoxArray and DistributionMapping and
-// fill with existing fine and coarse data.
-// overrides the pure virtual function in AmrCore
+/**
+ * Remake an existing level using provided BoxArray and DistributionMapping and
+ * fill with existing fine and coarse data.
+ * overrides the pure virtual function in AmrCore
+ * @param[in   ] lev     level to make
+ * @param[in   ] time    current time
+ * @param[in   ] ba      BoxArray for the level
+ * @param[in   ] dm      DistributionMapping for the level
+ */
 void
 REMORA::RemakeLevel (int lev, Real time, const BoxArray& ba, const DistributionMapping& dm)
 {
@@ -226,11 +239,18 @@ REMORA::RemakeLevel (int lev, Real time, const BoxArray& ba, const DistributionM
 #endif
 }
 
-// Make a new level from scratch using provided BoxArray and DistributionMapping.
-// This is called both for initialization and for restart
-// (overrides the pure virtual function in AmrCore)
-// main.cpp --> REMORA::InitData --> InitFromScratch --> MakeNewGrids --> MakeNewLevelFromScratch
-//                                         restart  --> MakeNewGrids --> MakeNewLevelFromScratch
+/**
+ * Make a new level from scratch using provided BoxArray and DistributionMapping.
+ * This is called both for initialization and for restart
+ * (overrides the pure virtual function in AmrCore)
+ * main.cpp --> REMORA::InitData --> InitFromScratch --> MakeNewGrids --> MakeNewLevelFromScratch
+ *                                        restart  --> MakeNewGrids --> MakeNewLevelFromScratch
+ *
+ * @param[in   ] lev     level to make
+ * @param[in   ] time    current time
+ * @param[in   ] ba      BoxArray for the level
+ * @param[in   ] dm      DistributionMapping for the level
+ */
 void REMORA::MakeNewLevelFromScratch (int lev, Real time, const BoxArray& ba,
                                      const DistributionMapping& dm)
 {
@@ -297,6 +317,9 @@ void REMORA::MakeNewLevelFromScratch (int lev, Real time, const BoxArray& ba,
 #endif
 }
 
+/**
[…]
