+++ Source/BoundaryConditions/REMORA_BoundaryConditions_cons.cpp
@@ -3,20 +3,22 @@
 
 using namespace amrex;
 
-//
-// mf is the multifab to be filled
-// icomp is the index into the MultiFab -- if cell-centered this can be any value
-//       from 0 to NCONS-1, if face-centered this must be 0
-// ncomp is the number of components -- if cell-centered (var_idx = 0) this can be any value
-//       from 1 to NCONS as long as icomp+ncomp <= NCONS-1.  If face-centered this
-//       must be 1
-// time is the time at which the data should be filled
-// bccomp is the index into both domain_bcs_type_bcr and bc_extdir_vals for icomp = 0  --
-//     so this follows the BCVars enum
-// n_not_fill perimeter of cells in x and y where BCs are not applied for conditions other than ext_dir.
-//     Foextrap is done based on the values at dom_lo-n_not_fill and dom_hi+n_not_fill. Reflecting conditions
-//     are untested.
-//
+/**
+ * @param[inout] dest_arr      data on which to apply BCs
+ * @param[in   ] bx            box to update on
+ * @param[in   ] valid_bx      valid box
+ * @param[in   ] domain        domain box
+ * @param[in   ] dxInv         pm or pn
+ * @param[in   ] mskr          land-sea mask on rho-points
+ * @param[in   ] msku          land-sea mask on u-points
+ * @param[in   ] mskv          land-sea mask on v-points
+ * @param[in   ] calc_arr      data to use in the RHS of calculations
+ * @param[in   ] icomp         component to update
+ * @param[in   ] ncomp         number of components to update, starting from icomp
+ * @param[in   ] time          current time
+ * @param[in   ] bccomp        index into both domain_bcs_type_bcr and bc_extdir_vals for icomp=0
+ * @param[in   ] n_not_fill    perimter of cells in x and y where BCs are not applied for non-ext_dir conditions
+ */
 void REMORAPhysBCFunct::impose_cons_bcs (const Array4<Real>& dest_arr, const Box& bx, const Box& valid_bx, const Box& domain,
                                         const GpuArray<Real,AMREX_SPACEDIM> /*dxInv*/, const Array4<const Real>& mskr,
                                         const Array4<const Real>& msku, const Array4<const Real>& mskv,
@@ -73,9 +75,7 @@ void REMORAPhysBCFunct::impose_cons_bcs (const Array4<Real>& dest_arr, const Box
                 if (bc_ptr[n].lo(0) == REMORABCType::ext_dir) {
                     dest_arr(i,j,k,icomp+n) = l_bc_extdir_vals_d[n][0] * mskr(i,j,0);
                 } else if (bc_ptr[n].lo(0) == REMORABCType::orlanski_rad) {
-                    Real grad_lo_im1   = (calc_arr(dom_lo.x-1,j  ,k,icomp+n) - calc_arr(dom_lo.x-1,j-1,k,icomp+n)) * mskv(i,j,0);
[…]

[…]
