### 2025-05-12 91a4b4e
Make docs workflow run both on push and pull_request
