+++ .github/workflows/docs.yml
@@ -1,6 +1,6 @@
 name: DocHTML -- Build and Deploy
 
-on: [pull_request]
+on: [push, pull_request]
 
 concurrency:
   group: ${{ github.ref }}-${{ github.head_ref }}-docs
