### 2025-05-13 f0da867
include netcdf in doxygen
