+++ Docs/Doxyfile.in
@@ -2013,7 +2013,7 @@ INCLUDE_FILE_PATTERNS  =
 # recursively expanded use the := operator instead of the = operator.
 # This tag requires that the tag ENABLE_PREPROCESSING is set to YES.
 
-PREDEFINED             = DIFFUSION BL_SPACEDIM=3 REMORA REMORA_USE_MOL
+PREDEFINED             = DIFFUSION BL_SPACEDIM=3 REMORA REMORA_USE_MOL REMORA_USE_NETCDF
 
 # If the MACRO_EXPANSION and EXPAND_ONLY_PREDEF tags are set to YES then this
 # tag can be used to specify a list of macro names that should be expanded. The
