### 2025-05-15 e7e985f
make sure MPI_Comm in amrex namespace to match AMReX changes
