+++ Source/IO/REMORA_console_io.cpp
@@ -22,7 +22,7 @@ const std::string dbl_line = std::string(78, '=') + "\n";
 const std::string dash_line = "\n" + std::string(78, '-') + "\n";
 } // namespace
 
-void REMORA::print_usage(MPI_Comm comm, std::ostream& out)
+void REMORA::print_usage(amrex::MPI_Comm comm, std::ostream& out)
 {
 #ifdef AMREX_USE_MPI
     int irank = 0;
@@ -47,7 +47,7 @@ Optional:
 )doc" << std::endl;
 }
 
-void REMORA::print_error(MPI_Comm comm, const std::string& msg)
+void REMORA::print_error(amrex::MPI_Comm comm, const std::string& msg)
 {
 #ifdef AMREX_USE_MPI
     int irank = 0;
@@ -64,7 +64,7 @@ void REMORA::print_error(MPI_Comm comm, const std::string& msg)
     std::cout << "ERROR: " << msg << std::endl;
 }
 
-void REMORA::print_banner(MPI_Comm comm, std::ostream& out)
+void REMORA::print_banner(amrex::MPI_Comm comm, std::ostream& out)
 {
 #ifdef AMREX_USE_MPI
     int irank = 0;
