### 2025-05-15 b61acf2
fix MPI_Comm issues for both MPI/non-MPI builds
