### 2025-05-16 4b6c557
paper updates for JOSS
