+++ paper/paper.bib
@@ -1,22 +1,22 @@
-@article{AMReX:JOSS, 
-         doi = {10.21105/joss.01370}, 
-         url = {https://doi.org/10.21105/joss.01370}, 
-         year = {2019}, 
-         publisher = {The Open Journal}, 
-         volume = {4}, 
-         number = {37}, 
-         pages = {1370}, 
-         author = {Zhang, W. and Almgren, A. and Beckner, V. and Bell, J. and Blaschke, J. and Chan, C. and Day, M. and Friesen, B. and Gott, K. and Graves, D. and Katz, M.P. and Myers, A. and Nguyen, T. and Nonaka, A. and Rosso, M. and Williams, S. and Zingale, M.}, 
-         title = {AMReX: a framework for block-structured adaptive mesh refinement}, 
-         journal = {Journal of Open Source Software} 
-} 
+@article{AMReX:JOSS,
+         doi = {10.21105/joss.01370},
+         url = {https://doi.org/10.21105/joss.01370},
+         year = {2019},
+         publisher = {The Open Journal},
+         volume = {4},
+         number = {37},
+         pages = {1370},
+         author = {Zhang, W. and Almgren, A. and Beckner, V. and Bell, J. and Blaschke, J. and Chan, C. and Day, M. and Friesen, B. and Gott, K. and Graves, D. and Katz, M.P. and Myers, A. and Nguyen, T. and Nonaka, A. and Rosso, M. and Williams, S. and Zingale, M.},
+         title = {AMReX: a framework for block-structured adaptive mesh refinement},
+         journal = {Journal of Open Source Software}
+}
 
-@article{AMReX:IJHPCA, 
+@article{AMReX:IJHPCA,
          doi = {10.1177/10943420211022811},
-         year = {2021}, 
-         volume = {35}, 
-         number = {6}, 
-         pages = {508-526}, 
+         year = {2021},
+         volume = {35},
+         number = {6},
+         pages = {508-526},
          author = {Zhang, W. and Myers, A. and Gott, K. and Almgren, A. and Bell, J.},
          title = {AMReX: Block-structured adaptive mesh refinement for multiphysics applications},
          journal = {International Journal of High Performance Computing Applications}
@@ -104,7 +104,7 @@
 @article{haidvogel.ea:08,
   doi = {10.1016/j.jcp.2007.06.016},
   title={Ocean forecasting in terrain-following coordinates: Formulation and skill assessment of the {R}egional {O}cean {M}odeling {S}ystem},
-  author = {D.B. Haidvogel and H. Arango and W.P. Budgell and B.D. Cornuelle and E. Curchitser and E. {Di Lorenzo} and K. Fennel and W.R. Geyer and A.J. Hermann and L. Lanerolle and J. Levin and J.C. McWilliams and A.J. Miller and A.M. Moore and T.M. Powell and A.F. Shchepetkin and C.R. Sherwood and R.P. Signell and J.C. Warner and J. Wilkin},  
+  author = {D.B. Haidvogel and H. Arango and W.P. Budgell and B.D. Cornuelle and E. Curchitser and E. {Di Lorenzo} and K. Fennel and W.R. Geyer and A.J. Hermann and L. Lanerolle and J. Levin and J.C. McWilliams and A.J. Miller and A.M. Moore and T.M. Powell and A.F. Shchepetkin and C.R. Sherwood and R.P. Signell and J.C. Warner and J. Wilkin},
   journal={Journal of Computational Physics},
   volume={227},
   number={7},
@@ -122,4 +122,52 @@
   number={2},
   pages={381--389},
   year={1995}
+}
+
+@article{fairall:96,
+  doi = {10.1029/95JC03205},
[…]
