+++ paper/paper.md
@@ -44,17 +44,17 @@ affiliations:
  - name: Argonne National Laboratory
    index: 3
 
-date: January 2025
+date: May 2025
 
 bibliography: paper.bib
 ---
 
 # Summary
 
-The Regional Model of the Ocean Refined Adaptively (REMORA) is a new implementation 
-of an existing community standard ocean model, 
-the Regional Ocean Modeling System 
-[ROMS, @haidvogel.ea:08; @shchepetkin.mcwilliams:05] that simulates estuarine and oceanic 
+The Regional Model of the Ocean Refined Adaptively (REMORA) is a new implementation
+of an existing community standard ocean model,
+the Regional Ocean Modeling System
+[ROMS, @haidvogel.ea:08; @shchepetkin.mcwilliams:05] that simulates estuarine and oceanic
 dynamics using the latest high-performance computing architectures.
 REMORA employs hierarchical parallelism using an MPI+X model, where X may be OpenMP on
 multicore CPU-only systems, or CUDA, HIP, or SYCL on GPU-accelerated systems.
@@ -62,7 +62,7 @@ It is able to be built and run in both single and double precision.
 REMORA is built on AMReX [@AMReX:JOSS; @AMReX:IJHPCA],
 a block-structured adaptive mesh refinement (AMR) software framework that
 provides the underlying performance-portable software infrastructure for block-structured mesh operations.
-REMORA, like ROMS, is a regional model, meaning that it is generally used for limited domains, 
+REMORA, like ROMS, is a regional model, meaning that it is generally used for limited domains,
 and as such requires boundary conditions derived analytically, or from larger-scale models.
 
 # Statement of need
@@ -70,50 +70,50 @@ and as such requires boundary conditions derived analytically, or from larger-sc
 Most widely used ocean modeling codes today do not have the
 ability to use GPU acceleration, which limits their ability to
 efficiently utilize current and next-generation high performance computing
-architectures.  REMORA provides an ocean modeling capability (based on ROMS, a proven FORTRAN code
-that runs efficiently on CPUs) that is able to run on all of the latest high-performance
+architectures. Oceananigans (Julia-based) [oceananigans] and Veros (Python-based) [veros] have both been developed as flexible,
+GPU-enabled models that can be used to simulate regional and global applications.
+REMORA is a C++-based alternative. Its ocean modeling capability is directly based on ROMS (a proven FORTRAN code
+that runs efficiently on CPUs), and is able to run on all of the latest high-performance
 computing architectures, from laptops to supercomputers, CPU-only or GPU-accelerated.
-In addition, REMORA is based on AMReX,
+REMORA is based on AMReX,
 a modern, well-supported adaptive mesh refinement (AMR) library,
 which provides a performance portable interface that shields REMORA
-from most of the detailed changes needed to adapt to new systems.
+from most of the detailed changes needed to adapt to new systems. It will also provide a straightforward
+avenue for incorporating AMR into REMORA.
 The active and large developer community contributing to AMReX helps ensure
 that REMORA will continue to run efficiently as architectures and operating systems
 evolve.
 
 # REMORA Features
 
-Like ROMS, REMORA:
+REMORA re-implements the core functionality of ROMS with the AMReX C++ framework.
+Unless otherwise indicated, the equations and solvers used are the same in REMORA and ROMS.
+REMORA solves the incompressible time-dependent Navier-Stokes equation with the Boussinesq
+and hydrostatic approximations [@haidvogel.ea:08; @shchepetkin.mcwilliams:05].
+The equations are solved on a curvilinear Arakawa C-grid with a stretched, terrain-following vertical s-coordinate.
[…]
