### 2025-05-16 e66b75c
fix spelling and bib
