+++ paper/paper.bib
@@ -137,7 +137,7 @@
 }
 
 @article{fairall:03,
-  doi={10.1175/1520-0442(2003)016<0571:BPOASF>2.0.CO;2}
+  doi={10.1175/1520-0442(2003)016<0571:BPOASF>2.0.CO;2},
   title={Bulk parameterization of air--sea fluxes: Updates and verification for the COARE algorithm},
   author={Fairall, Chris W and Bradley, Edward F and Hare, JE and Grachev, Andrey A and Edson, James B},
   journal={Journal of climate},
