### 2025-05-16 cc67bde
fix paper citations
