### 2025-05-17 a75d06b
fix citations again
