### 2025-05-17 seahorce-scidac/ROMSX#375 @hklion
@jwallwork23 Thanks for the note! I've updated the docstrings so doxygen (and comments) should be showing a lot more info now.
