### 2025-05-17 3bfb5fa
more citation formatting
