### 2025-05-17 #7958 @hklion
@jwallwork23 I've finished your recommended modifications to the paper and documentation. Please let me know if there are any additional comments.
