### 2025-05-19 #7958 @jatkinson1000
Hi @hklion @jwallwork23 @bastorer @esclapez, just checking in to see how this is going?

@jwallwork23 it looks like things are progressing nicely with your review, hopefully we can tick off the last few things this week.

@esclapez It looks like you have done the basic checks - might you be able to look at the software and paper this week?
@jwallwork23 has probably caught a few of the low-hanging issues already if you review the conversation above.

@bastorer please could you generate your checklist using `@editorialbot generate my checklist`?
It would be great to get your perspective on the submission from the domain side.
Alternatively please let me know if you feel you are no longer able to review this submission.

We are now past the planned schedule, so please could we try and get this done over the next week.
I will revisit this on Monday 26th.
