### 2025-05-21 4dfa68f
ocean_times is now a vector and no longer hardcoded size
