### 2025-05-27 #7958 @jwallwork23
> How do you feel about [@hklion](https://github.com/hklion)'s updates following your comments. Does this cover the open points on the checklist or are there any further things you feel need adding/changing.

@hklion @jatkinson1000 Thanks for the paper updates and reminder. I can confirm I'm happy with almost all of the checklist items. The one remaining is (again) for me to test on GPU. I had some CUDA version incompatibility issues on the system I was using but will try again on a different system.
