### 2025-05-27 seahorce-scidac/ROMSX#375 @joewallwork
Thanks for these updates - really useful! Happy to close this issue.
