### 2025-05-27 #7958 @jwallwork23
> The one remaining is (again) for me to test on GPU. I had some CUDA version incompatibility issues on the system I was using but will try again on a different system.

I managed to run the tests with CUDA enabled on an Ampere node. All the tests passed except the `BoundaryLayer` one.

```
13:             variable name            absolute error            relative error
13:                                         (||A - B||)         (||A - B||/||A||)
13:  ----------------------------------------------------------------------------
13:  level = 0
13:  temp                                0.000189528587           8.065450234e-06
13:  salt                               2.274342295e-05           7.163395573e-07
13:  x_velocity                         1.948696837e-07           4.850269304e-06
13:  y_velocity                         4.771215306e-08           2.319774173e-07
13:  z_velocity                                       0                         0
13/13 Test #13: BoundaryLayer ....................***Failed   13.47 sec
```

@hklion is there anything that I might need to configure differently for this test to pass? I requested 2 MPI ranks and 1 GPU for the job.

Build script

```sh
#!/bin/bash

BUILD_DIR="$(pwd)/build"
rm -rf "${BUILD_DIR}"
mkdir "${BUILD_DIR}"
cd "${BUILD_DIR}"

cmake .. \
        -DCMAKE_INSTALL_PREFIX:PATH="${BUILD_DIR}" \
        -DCMAKE_BUILD_TYPE=Release \
        -DCMAKE_CXX_COMPILER:STRING=mpicxx \
        -DREMORA_ENABLE_CUDA:BOOL=ON \
        -DREMORA_ENABLE_MPI:BOOL=ON \
        -DREMORA_ENABLE_FCOMPARE:BOOL=ON \
        -DREMORA_ENABLE_TESTS:BOOL=ON \
        -DREMORA_USE_CPP:BOOL=ON \
        -DAMReX_CUDA_ERROR_CROSS_EXECUTION_SPACE_CALL:BOOL=ON\
        -DAMReX_CUDA_ERROR_CAPTURE_THIS:BOOL=ON
make -j 4 USE_MPI=TRUE USE_OMP=FALSE USE_CUDA=TRUE
```

Run script

```sh
#!/bin/bash

cd build
ctest -V
```
