### 2025-05-31 44980fc
add a quickstart to docs and a few lines on amrvis
