### 2025-06-02 #7958 @jatkinson1000
Thanks for your service @esclapez!

It looks like this is close to acceptance.
@hklion did you have any advice for @jwallwork23's failing `BoundaryLayer` test?

I'll try and take an editorial look over the paper this week to speed things up once this has been resolved.
