### 2025-06-02 #7958 @hklion
@jatkinson1000 Sorry for not being clear earlier. I believe @esclapez and @jwallwork23 had the same `BoundaryLayer` test failure. So it should now be fixed for both of them.

@jwallwork23 If you still see a failure after the update, let me know and I'll dig in more.
