### 2025-06-02 d21c34e
start adding time controls for outputs
