+++ Source/IO/REMORA_ReadFromBdryNetcdf.cpp
@@ -94,11 +94,12 @@ read_bdry_from_netcdf (const Box& domain, const std::string& nc_bdry_file,
 
     // Even though we may not read in all the variables, we need to make the arrays big enough for them (for now)
 
+    int nt_prior = bdy_data_xlo.size();
     // Our outermost loop is time
-    bdy_data_xlo.resize(ntimes);
-    bdy_data_xhi.resize(ntimes);
-    bdy_data_ylo.resize(ntimes);
-    bdy_data_yhi.resize(ntimes);
+    bdy_data_xlo.resize(ntimes + nt_prior);
+    bdy_data_xhi.resize(ntimes + nt_prior);
+    bdy_data_ylo.resize(ntimes + nt_prior);
+    bdy_data_yhi.resize(ntimes + nt_prior);
 
     amrex::IntVect plo(lo);
     amrex::IntVect phi(hi);
@@ -225,34 +226,34 @@ read_bdry_from_netcdf (const Box& domain, const std::string& nc_bdry_file,
 
                 if        (bdyVarType == BdyVars::u) {
                     for (int nt(0); nt < ntimes; ++nt) {
-                        bdy_data_xlo[nt].push_back(FArrayBox(bx_for_u, 1, Arena_Used)); // u
+                        bdy_data_xlo[nt + nt_prior].push_back(FArrayBox(bx_for_u, 1, Arena_Used)); // u
                     }
                 } else if (bdyVarType == BdyVars::v) {
                     for (int nt(0); nt < ntimes; ++nt) {
-                        bdy_data_xlo[nt].push_back(FArrayBox(bx_for_v , 1, Arena_Used)); // v
+                        bdy_data_xlo[nt + nt_prior].push_back(FArrayBox(bx_for_v , 1, Arena_Used)); // v
                     }
                 } else if (bdyVarType == BdyVars::t) {
                     for (int nt(0); nt < ntimes; ++nt) {
-                        bdy_data_xlo[nt].push_back(FArrayBox(bx_for_t, 1, Arena_Used)); // temp
+                        bdy_data_xlo[nt + nt_prior].push_back(FArrayBox(bx_for_t, 1, Arena_Used)); // temp
                     }
                 } else if (bdyVarType == BdyVars::s) {
                     for (int nt(0); nt < ntimes; ++nt) {
-                        bdy_data_xlo[nt].push_back(FArrayBox(bx_for_t, 1, Arena_Used)); // salt
+                        bdy_data_xlo[nt + nt_prior].push_back(FArrayBox(bx_for_t, 1, Arena_Used)); // salt
                     }
                 } else if (bdyVarType == BdyVars::ubar) {
                     Box xlo_ubar(makeSlab(bx_for_u,2,0));
                     for (int nt(0); nt < ntimes; ++nt) {
-                        bdy_data_xlo[nt].push_back(FArrayBox(xlo_ubar, 1, Arena_Used)); // ubar
+                        bdy_data_xlo[nt + nt_prior].push_back(FArrayBox(xlo_ubar, 1, Arena_Used)); // ubar
                     }
                 } else if (bdyVarType == BdyVars::vbar) {
                     Box xlo_vbar(makeSlab(bx_for_v,2,0));
                     for (int nt(0); nt < ntimes; ++nt) {
-                        bdy_data_xlo[nt].push_back(FArrayBox(xlo_vbar, 1, Arena_Used)); // ubar
+                        bdy_data_xlo[nt + nt_prior].push_back(FArrayBox(xlo_vbar, 1, Arena_Used)); // ubar
                     }
                 } else if (bdyVarType == BdyVars::zeta) {
                     Box xlo_zeta(makeSlab(bx_for_t,2,0));
                     for (int nt(0); nt < ntimes; ++nt) {
[…]
