### 2025-06-04 d6ce4ba
add option to output plot and checkpoint files at simulation time intervals
