### 2025-06-05 #7958 @jwallwork23
> [@jwallwork23](https://github.com/jwallwork23) If you still see a failure after the update, let me know and I'll dig in more.

After updating and rebuilding I now get
```
100% tests passed, 0 tests failed out of 13
```
so all good! Thanks for this update @hklion.

With that I've now completed my review checklist. @jatkinson1000 the submission is good to go from my side.
