### 2025-06-09 #7958 @jatkinson1000
@hklion This is all looking like it is ready for acceptance from the review side.
Thank you @jwallwork23 and @esclapez for your service!

The next steps are to provide a tagged release for the version of the code that was reviewed, and place this same version in an online archive with a DOI (e.g. Zenodo, though other options are available).

Once you have done this please could you post the version number and archive DOI here please.

In the meantime I will take a proof-read of the paper to try and catch any small final issues from the editorial side.
