### 2025-06-09 7eee120
Typographical change Replace 'adaptive mesh refinement' with 'AMR'.
