### 2025-06-09 000b78e
Add citation of the Oceananigans software.
