### 2025-06-09 #7958 @hklion
@jatkinson1000 I merged your changes to the paper. Thank you!

The latest version of the code is [v1.3](https://github.com/seahorce-scidac/REMORA/releases/tag/v1.3).

Zenodo DOI: https://zenodo.org/records/15626120
