### 2025-06-10 #7958 @kthyng
@hklion Sorry I don't know off the top of my head nor with a quick search. The goal would be to get it near to your name under the software title on https://zenodo.org/records/15626120, in the same order as the paper author list.
